fabric: Add mr_count field to domain attribute
Fixes #3084

Allow the provider to indicate the number of memory registrations
that it can support on a domain.  This also allows the app a way
to indicate how many memory regions it may need.  Providers can
use this value to optimize caching of registrations.

Add missing output to fi_tostr as part of adding mr_cnt

Change-Id: Id1f6f17ad67c89741137a72cead380101ee56dc8
Signed-off-by: Sean Hefty <sean.hefty@intel.com>
Signed-off-by: Dmitry Gladkov <dmitry.gladkov@intel.com>