util/mr_cache: Add framework for memory registration cache
This is derived from work by Dmitry Gladkov, which was based on
the registration cache in the gni provider.

The interface for the cache is comprised of initialization and
cleanup routines, plus two calls: search and delete.  Search
will first search the cache for a region that contains the
provided input region.  If no existing region is found, the
new region is added to the cache.  Delete marks that the user
is done with the region.  Every search call should be paired
with a delete call.

If caching is enabled, the freeing of a delete region will be
deferred until it is both no longer being accessed, and is the
region that has the oldest access time.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>