util/pollfds: Add hot pollfd list
When connected to a large number of peers, the poll() system call
can be slow because of the number of fd's to check.  If only a
few peers are actively sending or receiving from the local
endpoint, we can reduce the overhead of checking by maintaining
an array of fd's which are active, or hot.

This adds a hot fd list to the pollfds abstraction, along with
a fairly simple mechanism for tracking fd's which are hot.  A
fairness counter is also maintained to ensure that the full fd
list is periodically checked.  The fairness counter will be
controllable though an environment variable, with a default
value set in follow on patches based on experimentation.

An fd is considered hot if it has data to send and the caller
is blocked waiting for a POLLOUT event.  It is also hot if any
data has been received on the fd.  A hit count is used track
activity, with the hit count reset every time the full fd list
is polled.  Hot fd's are added to a hot pollfd set.   If an
fd on the hot list has not received any events before the
fairness counter expires, it is removed from the list.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>